requirements.txt
python-crontab==3.0.0
Requests==2.31.0
pymysql==1.1.0
pydantic==2.5.3
python-dotenv==1.0.0
uvicorn==0.27.1
fastapi==0.109.2
jinja2==3.1.3
croniter==2.0.3